+++ lib/src/enums/platform.dart
@@ -1,0 +1,34 @@
+// ignore_for_file: constant_identifier_names
+
+/// The rich response message integration platform
+enum DialogPlatform {
+  /// Default platform
+  PLATFORM_UNSPECIFIED,
+
+  /// Facebook
+  FACEBOOK,
+
+  /// Slack
+  SLACK,
+
+  /// Telegram
+  TELEGRAM,
+
+  /// Kik
+  KIK,
+
+  /// Skype
+  SKYPE,
+
+  /// Line
+  LINE,
+
+  /// Viber
+  VIBER,
+
+  /// Google Assistant
+  ACTIONS_ON_GOOGLE,
+
+  /// Google Hangouts
+  GOOGLE_HANGOUTS,
+}